+++ src/Gravatar.php
@@ -1,24 +1,23 @@
 <?php
 /**
- * Gravatar plugin for Craft CMS 3.x
+ * Gravatar plugin for Craft CMS 4.x
  *
  * Adds gravatar support
  *
- * @link      https://github.com/noxify
- * @copyright Copyright (c) 2018 Marcus Reinhardt
+ * @link      https://github.com/BelniakMedia
+ * @copyright Copyright (c) 2022 Belniak Media, Inc.
  */
 
-namespace noxify\gravatar;
+namespace belniakmedia\gravatar;
 
-use noxify\gravatar\services\Url as UrlService;
-use noxify\gravatar\services\Img as ImgService;
-use noxify\gravatar\variables\GravatarVariable;
-use noxify\gravatar\models\Settings;
+use belniakmedia\gravatar\services\Url as UrlService;
+use belniakmedia\gravatar\services\Img as ImgService;
+use belniakmedia\gravatar\variables\GravatarVariable;
+use belniakmedia\gravatar\models\Settings;
 
 use Craft;
+use craft\base\Model;
 use craft\base\Plugin;
-use craft\services\Plugins;
-use craft\events\PluginEvent;
 use craft\web\twig\variables\CraftVariable;
 
 use yii\base\Event;
@@ -26,8 +25,8 @@ use yii\base\Event;
 /**
  * Class Gravatar
  *
- * @author    Marcus Reinhardt
- * @package   Gravatar
+ * @author    Belniak Media, Inc.
+ * @package   belniakmedia/craft-gravatar
  * @since     1.0.0
  *
  * @property  UrlService $url
@@ -35,24 +34,9 @@ use yii\base\Event;
  */
 class Gravatar extends Plugin
 {
-    // Static Properties
-    // =========================================================================
 
-    /**
-     * @var Gravatar
-     */
-    public static $plugin;
-
-    // Public Properties
-    // =========================================================================
-
-    /**
-     * @var string
-     */
-    public $schemaVersion = '1.0.0';
-
-    // Public Methods
-    // =========================================================================
+    public static Gravatar $plugin;
+    public string $schemaVersion = '1.0.0';
 
     /**
      * @inheritdoc
@@ -72,15 +56,6 @@ class Gravatar extends Plugin
             }
         );
 
-        Event::on(
-            Plugins::class,
-            Plugins::EVENT_AFTER_INSTALL_PLUGIN,
-            function (PluginEvent $event) {
-                if ($event->plugin === $this) {
-                }
-            }
-        );
-
         Craft::info(
             Craft::t(
                 'gravatar',
@@ -91,13 +66,10 @@ class Gravatar extends Plugin
         );
     }
 
-    // Protected Methods
-    // =========================================================================
-
     /**
      * @inheritdoc
      */
-    protected function createSettingsModel()
+    protected function createSettingsModel(): ?Model
     {
         return new Settings();
     }